pyproject.toml
[tool.poetry]
name = "scratch1"
version = "0.1.0"
description = ""
authors = ["Martín Marrari <mmarrari@een.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
